+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.2.14
+
+### Patch Changes
+
+-   f8b24a3a: fix additional msgs and readme
+-   Updated dependencies [f8b24a3a]
+    -   @sap-ux/inquirer-common@0.2.6
+
+## 0.2.13
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.2.5
+
+## 0.2.12
+
+### Patch Changes
+
+-   0d0d1784: fix prompt choice issue
+
 ## 0.2.11
 
 ### Patch Changes